--- backend/games/core/models/__init__.py
@@ -1,3 +1,0 @@
-from .product import Product
-from .game_dlc_link import GameDlcLink
-from .system_requirement import SystemRequirement
